backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
python-slugify
python-multipart
pytest
black
flake8
alembic
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic-settings>=2.0.3